+++ zipline_app/models/zipline_app/asset.py
@@ -11,9 +11,10 @@ class Asset(models.Model):
     asset_exchange = models.CharField(max_length=200)
     asset_name = models.CharField(max_length=200)
     asset_isin = models.CharField(max_length=20, null=True)
+    asset_currency = models.CharField(max_length=20, null=True)
 
     def __str__(self):
-        return "%s: %s" % (self.asset_symbol, self.asset_name)
+        return "%s: %s (%s)" % (self.asset_symbol, self.asset_name, self.asset_currency)
 
     def str(self):
       return self.__str__()
@@ -23,7 +24,8 @@ class Asset(models.Model):
           "symbol": self.asset_symbol,
           "exchange": self.asset_exchange,
           "name": self.asset_name,
-          "isin": self.asset_isin
+          "isin": self.asset_isin,
+          "currency": self.asset_currency
         }
 
     def delete(self):